+++ internal/database/queries.go
@@ -708,7 +708,7 @@ func (db *Database) InsertUserPermission(up *models.UserPermission) error {
 
 const query_GetUserPermissions = `SELECT id, user_id, permission, granted_at FROM user_permissions WHERE user_id = ?`
 
-func (db *Database) GetUserPermissions(userID int) ([]*models.UserPermission, error) {
+func (db *Database) GetUserPermissions(userID int64) ([]*models.UserPermission, error) {
 	rows, err := db.mainDB.Query(query_GetUserPermissions, userID)
 	if err != nil {
 		return nil, err
